requirements.txt
amqp==5.3.1
anyio==4.0.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.3
beautifulsoup4==4.13.3
billiard==4.2.4
celery==5.6.2
certifi==2023.7.22
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
cryptography==44.0.0
cssselect2==0.8.0
dj-database-url==2.2.0
Django==4.2.5
django-celery-beat==2.8.1
django-cors-headers==4.2.0
django-crispy-forms==2.4
django-dbbackup==4.1.0
django-filter==23.3
django-mathfilters==1.0.0
django-multiselectfield==0.1.12
django-redis==5.2.0
django-timezone-field==7.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-history==1.0.1
drf-spectacular==0.29.0
et-xmlfile==1.1.0
exceptiongroup==1.3.1
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
httpcore==0.18.0
httpx==0.25.0
idna==3.4
kombu==5.6.2
lxml==6.0.2
numpy==2.0.2
openpyxl==3.1.2
oscrypto==1.3.0
packaging==26.0
pandas==2.2.3
pillow==11.1.0
prompt_toolkit==3.0.52
psycopg2==2.9.10
psycopg2-binary==2.9.8
pycparser==2.22
pyHanko==0.33.0
pyhanko-certvalidator==0.29.1
PyJWT==2.8.0
pypdf==6.7.4
PyPDF2==3.0.1
python-bidi==0.4.2
python-crontab==3.3.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.7
reportlab==3.6.13
requests==2.32.5
six==1.16.0
sniffio==1.3.0
soupsieve==2.6
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.4.0
typing_extensions==4.8.0
tzdata==2025.3
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.6.0
webencodings==0.5.1
whitenoise==6.2.0
xhtml2pdf==0.2.11